+++ macros/src/derive/ref_type.rs
@@ -1,5 +1,5 @@
 use super::{EnumData, FieldData, StructData, VariantData};
-use crate::util;
+use aldrin_codegen::rust::names;
 use proc_macro2::{Span, TokenStream};
 use quote::{quote, ToTokens};
 use std::ptr;
@@ -166,7 +166,12 @@ impl VariantData<'_> {
 
     fn ctor(&self, enum_name: &Ident, variants: &[Self]) -> TokenStream {
         let name = self.name();
-        let ctor = Ident::new_raw(&util::to_snake_case(&name.unraw().to_string()), name.span());
+
+        let ctor = Ident::new_raw(
+            &names::enum_ref_type_ctor(&name.unraw().to_string()),
+            name.span(),
+        );
+
         let doc = self.doc();
 
         let ty_generics = variants.iter().filter_map(|var| {