+++ codegen/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support newtypes as the key type in sets and maps.
 - Doc comments are now copied into the generated code from the schema.
 - Support attributes for inline structs and enums.
+- Add the module `rust::names` with functions, that return various derived names.
 
 ### Changed